make discover screen reactive to PLATFORM_EXPERIMENTAL_APPS flag
